requirements.txt
langchain
gradio
langchain-openai
python-dotenv
langchain-community
langchain-core
pypdf